+++ config/config.php
@@ -43,6 +43,8 @@ return (new ConfigAggregator\ConfigAggregator([
     $isTestEnv
         ? new ConfigAggregator\PhpFileProvider('config/test/*.global.php')
         : new ConfigAggregator\ArrayProvider([]),
+    // Routes have to be loaded last
+    new ConfigAggregator\PhpFileProvider('config/autoload/routes.config.php'),
 ], 'data/cache/app_config.php', [
     Core\Config\BasePathPrefixer::class,
 ]))->getMergedConfig();